+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+
 - Updated task plan and task index (main)
 - Completed task `phase-1` (main)
 - Added 2 specifications (main)
@@ -27,4 +28,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated global project rules (main)
 - Completed task `phase-3` (main)
 - Added 5 specifications (main)
+- Completed task `phase-4` (main)